feat(client/shared): optimized shared components and made mcp editor shared comp
