+++ libs/shared/src/components/mcp-json-editor/index.ts
@@ -1,6 +1,3 @@
 export { EditorHeader } from "./editor-header";
 export { FunctionCard } from "./function-card";
-export { JsonTextarea } from "./json-textarea";
-export { MetaDataSection } from "./meta-data-section";
-export { SearchBar } from "./search-bar";
 export * from "./types";
