+++ packages/client/src/components/blocks-workspace/BlocksWorkspace.tsx
@@ -22,8 +22,8 @@ import {
 	WorkspaceManager,
 } from "../../components/workspace";
 import { DesignerContext } from "../../contexts";
+import { MCPJsonEditor } from "../shared";
 import { GraphPanel } from "../workspace/panels/GraphPanel";
-import { MCPJsonEditor } from "../workspace/panels/mcp-json-editor";
 import { BlocksWorkspaceActions } from "./BlocksWorkspaceActions";
 import { BlocksWorkspaceDev } from "./BlocksWorkspaceDev";
 import { DEFAULT_MENU } from "./menus/default-menu";
