+++ packages/client/src/components/shared/index.ts
@@ -1,10 +1,10 @@
+import { MCPJsonEditor } from "./mcp-json-editor";
 import { NavbarHeader } from "./NavbarHeader";
 import { NavbarLeft } from "./NavbarLeft";
 import { NavbarRight } from "./NavbarRight";
 import { Page } from "./Page";
 import { PlatformMessages } from "./PlatformMessages";
 import { PlatformSearch } from "./platform-search";
-
 export {
 	Page,
 	PlatformMessages,
@@ -12,4 +12,5 @@ export {
 	NavbarLeft,
 	NavbarRight,
 	NavbarHeader,
+	MCPJsonEditor,
 };
