+++ packages/client/src/index.css
@@ -2,4 +2,3 @@
 @import "@semoss/shared/globals.css";
 
 @source "./**/*.{ts,tsx}";
-@source "../../../libs/shared/src/**/*.{ts,tsx}";
